// src/app/main/apps/procesos/ascenso/models/ascenso.model.ts
export class EtapaResponseModel {
    idEtapa: number;
    idProceso: number;
    anio: number;
    idEtapaFase: number;
    codigoProceso: string;
    codigoEtapa: string;
    descripcionEtapa: string;
    descripcionRegimenLaboral: string;
    abreviaturaRegimenLaboral: string;
    regimenLaboral: string;
    descripcionEtapaFase: string;
    descripcionNumeroConvocatoria: string;
    descripcionTipoProceso: string;
    descripcionProceso: string;
    descripcionCausal: string;
    descripcionConvocatoria: string;
    fechaRegistro: string;
    fecha: string;
    descripcionGrupoModalidad: string;
    fechaRegistroEtapa: string;
    fechaRegistroProceso: string;
    idModalidad: string;
    idModalidadEducativa: string;
    idNivelEducativo: string;
    esEstadoEnProceso: boolean;
    abreviaturaModalidadEducativa: string;
    codigo: string;
    fechaCreacionProceso: string;
    tipoProceso: string;
    proceso: string;
    idCatalogoItem: string;
    codigoCatalogoItem: string;
    etapa: string;

}